Open agent terminals in dashboard shell
